+++ android_auto_huashan_aosprro60.sh
@@ -17,17 +17,14 @@ if [[ ! "$BuildMode" =~ "test" && ! "$BuildMode" =~ "nosync" ]]; then
   echo " [ Syncing $PhoneName repositories ]";
   echo "";
   cd $AndroidDir/;
-  repo forall -c 'echo "Cleaning project ${REPO_PROJECT}"; \
-                  git rebase --abort >/dev/null 2>&1; \
-                  git stash -u >/dev/null 2>&1; \
-                  git reset --hard HEAD >/dev/null 2>&1;';
+  if [[ ! "$BuildMode" =~ "unsafe" ]]; then
+    repo forall -c 'echo "Cleaning project ${REPO_PROJECT}"; \
+                    git rebase --abort >/dev/null 2>&1; \
+                    git stash -u >/dev/null 2>&1; \
+                    git reset --hard HEAD >/dev/null 2>&1;';
+  fi;
   repo sync -j8 --current-branch --detach -f --force-broken --force-sync -c --no-clone-bundle --no-tags;
   source ./build/envsetup.sh;
-
-  croot;
-  cd ./hardware/qcom/audio-caf/msm8960/; echo "";
-  git fetch https://github.com/arco/android_hardware_qcom_audio.git cm-13.0-caf-8960;
-  git reset --hard FETCH_HEAD;
 fi;
 
 
@@ -56,10 +53,14 @@ BuildSuccess="";
 if [[ ! "$BuildMode" =~ "synconly" ]]; then
   cd $ScriptsDir/;
   android_selection;
-  source $ScriptsDir/android_brunch.sh "automatic,$BuildMode,aosprro";
+  if [[ "$BuildMode" =~ "kernel" ]]; then
+    source $ScriptsDir/android_make_kernel.sh "release,otapackage" "AOSP-RRO-6.0.1-";
+  else
+    source $ScriptsDir/android_brunch.sh "automatic,$BuildMode,aosprro";
+  fi;
 
   # ROM Successful
-  if [ -f "$AndroidResult" ]; then
+  if [ ! -z "$AndroidResult" ] && [ -f "$AndroidResult" ]; then
     BuildSuccess="true";
   fi;
